+++ wicket/src/screens/component.rs
@@ -9,16 +9,13 @@ use crate::defaults::colors::*;
 use crate::defaults::dimensions::RectExt;
 use crate::defaults::dimensions::MENUBAR_HEIGHT;
 use crate::defaults::style;
+use crate::screens::ScreenId;
 use crate::widgets::Control;
 use crate::widgets::ControlId;
 use crate::widgets::HelpMenuState;
 use crate::widgets::{HelpButton, HelpButtonState, HelpMenu};
 use crate::widgets::{ScreenButton, ScreenButtonState};
-use crate::Action;
-use crate::Frame;
-use crate::ScreenEvent;
-use crate::ScreenId;
-use crate::State;
+use crate::wizard::{Action, Frame, ScreenEvent, State, Term};
 use crossterm::event::Event as TermEvent;
 use crossterm::event::{
     KeyCode, KeyEvent, KeyModifiers, MouseButton, MouseEvent, MouseEventKind,
@@ -303,11 +300,7 @@ impl ComponentScreen {
 }
 
 impl Screen for ComponentScreen {
-    fn draw(
-        &self,
-        state: &State,
-        terminal: &mut crate::Term,
-    ) -> anyhow::Result<()> {
+    fn draw(&self, state: &State, terminal: &mut Term) -> anyhow::Result<()> {
         terminal.draw(|f| {
             self.draw_background(f);
             self.draw_menubar(f, state);
